refactor: stage 2 — decompose loader into discover + transform + coordinator
- Create discover.lua for spec file scanning and loading
- Create transform.lua with pipeline-based lazy.nvim transformation
- Slim loader.lua to thin coordinator (328 -> 30 lines)
- Update core/init.lua to use discover for spec iteration
- Update debug.lua to use paths.is_dir helpers
