+++ lua/core/lib/loader.lua
@@ -1,328 +1,30 @@
-local debug_mod = require("core.lib.debug")
-local validate = require("core.lib.validate")
-local conditions = require("core.registry.conditions")
-local paths = require("core.lib.paths")
-local platform = require("core.lib.platform")
-local notify = require("core.lib.notify")
+local discover = require("core.lib.discover")
+local transform = require("core.lib.transform")
+local diagnostics_mod = require("core.lib.diagnostics")
 
 local M = {}
 
-M._specs = {}
-M._specs_by_name = {}
-M._errors = {}
-M._warnings = {}
-
-local function add_error(file, message, level)
-  table.insert(M._errors, { level = level or "critical", file = file, message = message })
-end
-
-local function add_warning(file, message)
-  table.insert(M._warnings, { level = "warning", file = file, message = message })
-end
-
-local function collect_results(file, errors, warnings)
-  for _, e in ipairs(errors) do
-    add_error(file, e.message, e.level)
-  end
-  for _, w in ipairs(warnings) do
-    add_warning(file, w.message)
-  end
-end
-
-local passthrough_fields = { "event", "cmd", "ft", "keys" }
-
-function M.get_plugin_dirs()
-  local root = debug_mod.get_luxvim_root()
-  local plugins_dir = paths.join(root, "lua", "plugins")
-
-  local entries = paths.scandir(plugins_dir, function(_, entry_type)
-    return entry_type == "directory"
-  end)
-
-  if #entries == 0 then
-    add_error("core.lib.loader",
-      "Plugin directory not found: " .. plugins_dir ..
-      "\nLuxVim root detected as: " .. root ..
-      "\nLaunch LuxVim from its directory or check installation.")
-  end
-
-  local dirs = {}
-  for _, entry in ipairs(entries) do
-    table.insert(dirs, { name = entry.name, path = paths.join(plugins_dir, entry.name) })
-  end
-  return dirs
-end
-
-function M.load_category_defaults(category_path)
-  local defaults_path = paths.join(category_path, "_defaults.lua")
-  local stat = vim.uv.fs_stat(defaults_path)
-  if not stat then
-    return {}
-  end
-
-  local ok, defaults = pcall(dofile, defaults_path)
-  if ok then
-    return defaults
-  end
-  return {}
-end
-
-function M.load_plugin_specs(category_path, category_name, defaults)
-  local specs = {}
-
-  local entries = paths.scandir(category_path, function(name, entry_type)
-    return entry_type == "file" and name:match("%.lua$") and name ~= "_defaults.lua"
-  end)
-
-  for _, entry in ipairs(entries) do
-    local file_path = paths.join(category_path, entry.name)
-    local ok, spec = pcall(dofile, file_path)
-
-    if not ok then
-      add_error(file_path, "failed to load: " .. tostring(spec))
-    elseif type(spec) ~= "table" then
-      add_error(file_path, "spec must be a table, got " .. type(spec))
-    else
-      local errors, warnings = validate.validate_plugin_spec(spec, file_path)
-      collect_results(file_path, errors, warnings)
-
-      if #errors == 0 then
-        spec._file = file_path
-        spec._category = category_name
-        spec = vim.tbl_deep_extend("keep", spec, defaults)
-        table.insert(specs, spec)
-      end
-    end
-  end
-
-  return specs
-end
-
-local function safe_eval(fn)
-  local ok, result = pcall(fn)
-  return ok and result
-end
-
-function M.evaluate_condition(cond)
-  if cond == nil then
-    return true
-  end
-
-  if type(cond) == "function" then
-    return safe_eval(cond)
-  end
-
-  if type(cond) == "string" then
-    local condition_fn = conditions[cond]
-    if condition_fn then
-      return safe_eval(condition_fn)
-    end
-    return false
-  end
-
-  return true
-end
-
-function M.transform_to_lazy(spec)
-  if not M.evaluate_condition(spec.cond) then
-    return nil
-  end
-
-  if spec.enabled == false then
-    return nil
-  end
-
-  local debug_name = debug_mod.resolve_debug_name(spec)
-  local use_debug = debug_mod.has_debug_plugin(debug_name)
-
-  local lazy_spec = {}
-
-  if use_debug then
-    lazy_spec.dir = debug_mod.get_debug_path(debug_name)
-    lazy_spec.name = debug_name .. "-debug"
-  else
-    lazy_spec[1] = spec.source
-  end
-
-  if spec.opts then
-    lazy_spec.opts = spec.opts
-  end
-
-  if spec.config ~= nil then
-    lazy_spec.config = spec.config
-  elseif spec.opts then
-    lazy_spec.config = true
-  end
-
-  if spec.dependencies then
-    lazy_spec.dependencies = M.resolve_dependencies(spec.dependencies)
-  end
-
-  for _, field in ipairs(passthrough_fields) do
-    if spec[field] then
-      lazy_spec[field] = spec[field]
-    end
-  end
-
-  if spec.build then
-    lazy_spec.build = M.transform_build(spec.build)
-  end
-
-  if spec.lazy then
-    if type(spec.lazy) == "table" then
-      lazy_spec = vim.tbl_deep_extend("force", lazy_spec, spec.lazy)
-    elseif spec.lazy == true then
-      lazy_spec.lazy = true
-    end
-  end
-
-  if spec.globals then
-    local lazy_init = lazy_spec.init
-    lazy_spec.init = function()
-      for key, value in pairs(spec.globals) do
-        vim.g[key] = value
-      end
-      if lazy_init then
-        lazy_init()
-      end
-    end
-  end
-
-  M._specs_by_name[debug_name] = spec
-
-  return lazy_spec
-end
-
-function M.resolve_dependencies(deps)
-  local resolved = {}
-  for _, dep in ipairs(deps) do
-    if type(dep) == "string" then
-      if M._specs_by_name[dep] then
-        local dep_spec = M._specs_by_name[dep]
-        local lazy_dep = M.transform_to_lazy(dep_spec)
-        if lazy_dep then
-          table.insert(resolved, lazy_dep)
-        end
-      else
-        table.insert(resolved, dep)
-      end
-    elseif type(dep) == "table" then
-      table.insert(resolved, dep)
-    end
-  end
-  return resolved
-end
-
-function M.transform_build(build)
-  if type(build) == "string" then
-    return build
-  end
-
-  if type(build) == "table" then
-    local cmd = build.cmd
-    if build.platforms then
-      if build.platforms[platform.os] then
-        cmd = build.platforms[platform.os]
-      end
-    end
-
-    if build.requires then
-      for _, exe in ipairs(build.requires) do
-        if vim.fn.executable(exe) ~= 1 then
-          if build.on_fail == "error" then
-            error("Build requires " .. exe .. " but it's not available")
-          elseif build.on_fail ~= "ignore" then
-            notify.warn("Build skipped: missing " .. exe)
-          end
-          return nil
-        end
-      end
-    end
-
-    if build.cond then
-      if not M.evaluate_condition(build.cond) then
-        return nil
-      end
-    end
-
-    return cmd
-  end
-
-  return nil
-end
+M._diagnostics = nil
 
 function M.discover_all()
-  M._specs = {}
-  M._specs_by_name = {}
-  M._errors = {}
-  M._warnings = {}
-
-  local dirs = M.get_plugin_dirs()
-
-  for _, dir in ipairs(dirs) do
-    local defaults = M.load_category_defaults(dir.path)
-    local specs = M.load_plugin_specs(dir.path, dir.name, defaults)
-
-    for _, spec in ipairs(specs) do
-      table.insert(M._specs, spec)
-      local name = debug_mod.resolve_debug_name(spec)
-      M._specs_by_name[name] = spec
-    end
-  end
-
-  return M._specs
+  M._diagnostics = diagnostics_mod.create()
+  discover.discover_all(M._diagnostics)
 end
 
 function M.get_lazy_specs()
-  local lazy_specs = {}
-
-  for _, spec in ipairs(M._specs) do
-    local lazy_spec = M.transform_to_lazy(spec)
-    if lazy_spec then
-      table.insert(lazy_specs, lazy_spec)
-    end
-  end
-
-  return lazy_specs
+  return transform.transform_all(discover._specs, discover._specs_by_name)
 end
 
 function M.report_errors()
-  local critical = vim.tbl_filter(function(e)
-    return e.level == "critical"
-  end, M._errors)
-
-  if #critical > 0 then
-    local msg = "[LuxVim] FATAL: Cannot start\n"
-    for _, e in ipairs(critical) do
-      msg = msg .. "  " .. e.file .. ": " .. e.message .. "\n"
-    end
-    vim.api.nvim_echo({ { msg, "ErrorMsg" } }, true, {})
-    return false
-  end
-
-  local non_critical = vim.tbl_filter(function(e)
-    return e.level ~= "critical"
-  end, M._errors)
-
-  for _, e in ipairs(non_critical) do
-    notify.warn("Plugin skipped: " .. e.file .. "\n  " .. e.message)
-  end
-
-  if #M._warnings > 0 then
-    vim.defer_fn(function()
-      notify.info("Started with " .. #M._warnings .. " warnings. Run :LuxVimErrors for details.")
-    end, 100)
-  end
-
-  return true
+  return M._diagnostics:report()
 end
 
 function M.get_errors()
-  return M._errors
+  return M._diagnostics and M._diagnostics.errors or {}
 end
 
 function M.get_warnings()
-  return M._warnings
+  return M._diagnostics and M._diagnostics.warnings or {}
 end
 
 return M
